Made image mocks a session-wide pytest fixture.
